% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smooth_log_periodicity}
\alias{smooth_log_periodicity}
\title{Smoothed log periodicity}
\usage{
smooth_log_periodicity(chord, d = 0.011)
}
\arguments{
\item{chord}{(numeric vector):
The chord's constituent notes, expressed on a MIDI scale.}

\item{d}{(numeric scalar):
Maximal allowed error in the algorithm's
interval approximation step, expressed as
a fraction of the original interval.
The default value, 0.011, corresponds to 'Rational Tuning II'
in Stolzenburg's paper.}
}
\value{
A numeric scalar identifying the chord's periodicity.
High values mean a higher period length, lower periodicity,
and lower consonance.
}
\description{
This function computes a chord's smoothed logarithmic periodicity,
after \insertCite{Stolzenburg2015;textual}{stolz15}.
}
\references{
\insertAllCited{}
}
